# docs/API_SPEC.yaml
openapi: 3.0.3
info:
  title: 알림 발송 시스템 API
  version: 0.0.1-SNAPSHOT
  description: |
    라이브클래스 백엔드 셀 채용 과제 C — 알림 발송 시스템.
    인증은 `X-User-Id` 헤더 기반의 단순 식별 (과제 허용 범위).

servers:
  - url: http://localhost:8080
    description: 로컬

paths:
  /api/v1/notifications:
    post:
      summary: 알림 발송 요청
      description: |
        알림 + Outbox 적재. 즉시 ACK 후 워커가 비동기 처리.
        `Idempotency-Key` 또는 `(recipientId, type, refType, refId)` 자연 키가 같으면 기존 알림을 그대로 반환 (200 OK).
        신규 적재 시 201 Created.
        `scheduledAt` 이 있으면 해당 시각부터 워커 claim 대상이 된다.
      parameters:
        - in: header
          name: X-User-Id
          required: true
          schema: { type: string }
        - in: header
          name: Idempotency-Key
          required: false
          schema: { type: string }
          description: 클라이언트 재시도 안전을 위한 멱등성 키.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNotificationRequest'
      responses:
        '201': { description: 신규 적재됨, content: { application/json: { schema: { $ref: '#/components/schemas/NotificationResponse' } } } }
        '200': { description: 멱등성/자연 키 중복으로 기존 알림 반환, content: { application/json: { schema: { $ref: '#/components/schemas/NotificationResponse' } } } }
        '400': { description: 입력 검증 실패, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }

    get:
      summary: 본인 알림 목록 조회
      parameters:
        - in: header
          name: X-User-Id
          required: true
          schema: { type: string }
        - in: query
          name: recipientId
          required: false
          schema: { type: string }
          description: 미지정 시 본인. 본인 외 지정 시 403 Forbidden.
        - in: query
          name: unreadOnly
          schema: { type: boolean, default: false }
        - in: query
          name: limit
          schema: { type: integer, default: 20, minimum: 1, maximum: 200 }
        - in: query
          name: offset
          schema: { type: integer, default: 0, minimum: 0 }
      responses:
        '200':
          description: 최신 createdAt 내림차순
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/NotificationResponse' }
        '400': { description: limit/offset 검증 실패 (BAD_REQUEST), content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '403': { description: 본인 외 사용자의 목록 조회 시도 (RECIPIENT_FORBIDDEN), content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }

  /api/v1/notifications/{id}:
    get:
      summary: 단건 조회
      parameters:
        - in: header
          name: X-User-Id
          required: true
          schema: { type: string }
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200': { description: 조회 성공, content: { application/json: { schema: { $ref: '#/components/schemas/NotificationResponse' } } } }
        '403': { description: 본인 외 사용자 시도 (NOTIFICATION_ACCESS_DENIED), content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: 존재하지 않음 (NOTIFICATION_NOT_FOUND), content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }

  /api/v1/notifications/{id}/retry:
    post:
      summary: DEAD_LETTER 알림 수동 재시도 (운영자 전용)
      description: |
        DEAD_LETTER 격리된 알림을 운영자(OPERATOR) 가 명시적으로 재시도.
        attempt_count 가 0 으로 리셋되어 다시 5회까지 자동 재시도됨.

        **권한**: `X-User-Role: OPERATOR` 헤더 필요.
        일반 USER 는 본인 알림이라도 직접 재시도 불가 (403 OPERATOR_ONLY).
      parameters:
        - in: header
          name: X-User-Id
          required: true
          schema: { type: string }
          description: 요청자(운영자) ID
        - in: header
          name: X-User-Role
          required: false
          schema:
            type: string
            enum: [USER, OPERATOR]
            default: USER
          description: 운영자 권한 필요. 누락 또는 OPERATOR 외 값이면 USER 로 취급되어 403.
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200': { description: PENDING 으로 복귀, content: { application/json: { schema: { $ref: '#/components/schemas/NotificationResponse' } } } }
        '403': { description: OPERATOR 권한 부재 (OPERATOR_ONLY), content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: 존재하지 않는 알림 (NOTIFICATION_NOT_FOUND), content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '409': { description: DEAD_LETTER 가 아닌 상태에서 호출 (INVALID_STATE) — 도메인 상태 위반, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }

  /api/v1/notifications/{id}/read:
    post:
      summary: 읽음 처리 (멀티 디바이스 멱등)
      description: 첫 호출은 readAt 을 set, 이후 호출은 no-op (newlyRead=false 반환).
      parameters:
        - in: header
          name: X-User-Id
          required: true
          schema: { type: string }
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: 처리 결과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadNotificationResponse'
        '403': { description: 본인 외 사용자 시도 (NOTIFICATION_ACCESS_DENIED), content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: 존재하지 않음 (NOTIFICATION_NOT_FOUND), content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }

  /api/v1/notification-templates:
    get:
      summary: 알림 템플릿 목록 조회
      responses:
        '200':
          description: 타입/채널별 템플릿 목록
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/NotificationTemplateResponse' }

  /api/v1/notification-templates/{type}/{channel}:
    get:
      summary: 알림 템플릿 단건 조회
      parameters:
        - in: path
          name: type
          required: true
          schema:
            type: string
            enum: [ENROLL_COMPLETED, PAYMENT_CONFIRMED, COURSE_STARTING_TOMORROW, ENROLL_CANCELLED]
        - in: path
          name: channel
          required: true
          schema:
            type: string
            enum: [EMAIL, IN_APP]
      responses:
        '200': { description: 조회 성공, content: { application/json: { schema: { $ref: '#/components/schemas/NotificationTemplateResponse' } } } }
        '404': { description: 템플릿 없음 (NOTIFICATION_TEMPLATE_NOT_FOUND), content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }

    put:
      summary: 알림 템플릿 수정 또는 생성 (운영자 전용)
      parameters:
        - in: header
          name: X-User-Id
          required: true
          schema: { type: string }
        - in: header
          name: X-User-Role
          required: false
          schema:
            type: string
            enum: [USER, OPERATOR]
            default: USER
        - in: path
          name: type
          required: true
          schema:
            type: string
            enum: [ENROLL_COMPLETED, PAYMENT_CONFIRMED, COURSE_STARTING_TOMORROW, ENROLL_CANCELLED]
        - in: path
          name: channel
          required: true
          schema:
            type: string
            enum: [EMAIL, IN_APP]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNotificationTemplateRequest'
      responses:
        '200': { description: 수정 또는 생성됨, content: { application/json: { schema: { $ref: '#/components/schemas/NotificationTemplateResponse' } } } }
        '400': { description: 입력 검증 실패, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '403': { description: OPERATOR 권한 부재 (OPERATOR_ONLY), content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }

  components:
  schemas:
    CreateNotificationRequest:
      type: object
      required: [recipientId, type, channel]
      properties:
        recipientId:
          type: string
          example: user-123
        type:
          type: string
          enum: [ENROLL_COMPLETED, PAYMENT_CONFIRMED, COURSE_STARTING_TOMORROW, ENROLL_CANCELLED]
        channel:
          type: string
          enum: [EMAIL, IN_APP]
        payload:
          type: object
          additionalProperties: true
          example: { title: "수강 신청 완료", body: "Spring Boot 입문" }
        refType:
          type: string
          example: ENROLLMENT
        refId:
          type: string
          example: "100"
        scheduledAt:
          type: string
          format: date-time
          nullable: true
          description: 지정 시 해당 시각부터 발송 대상이 된다. 미지정 시 즉시 발송 대상.

    NotificationResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        recipientId: { type: string }
        type: { type: string }
        channel: { type: string }
        payload: { type: object, additionalProperties: true }
        refType: { type: string, nullable: true }
        refId: { type: string, nullable: true }
        status:
          type: string
          enum: [PENDING, SENT, DEAD_LETTER]
        readAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        scheduledAt: { type: string, format: date-time, nullable: true }
        sentAt: { type: string, format: date-time, nullable: true }

    ReadNotificationResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        readAt: { type: string, format: date-time }
        newlyRead: { type: boolean }

    NotificationTemplateResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        type:
          type: string
          enum: [ENROLL_COMPLETED, PAYMENT_CONFIRMED, COURSE_STARTING_TOMORROW, ENROLL_CANCELLED]
        channel:
          type: string
          enum: [EMAIL, IN_APP]
        subjectTemplate: { type: string, example: "수강 신청 완료: {{courseName}}" }
        bodyTemplate: { type: string, example: "{{recipientName}}님, {{courseName}} 수강 신청이 완료되었습니다." }
        updatedAt: { type: string, format: date-time }

    UpdateNotificationTemplateRequest:
      type: object
      required: [subjectTemplate, bodyTemplate]
      properties:
        subjectTemplate: { type: string }
        bodyTemplate: { type: string }

    ErrorResponse:
      type: object
      description: |
        표준 오류 응답.
        code 는 안정적 식별자로 클라이언트의 분기 / retry 정책 결정에 사용 가능.
      properties:
        code:
          type: string
          enum:
            - NOTIFICATION_NOT_FOUND        # 404 — 존재하지 않는 알림
            - NOTIFICATION_TEMPLATE_NOT_FOUND # 404 — 존재하지 않는 템플릿
            - NOTIFICATION_ACCESS_DENIED    # 403 — 단일 알림 권한 거부
            - RECIPIENT_FORBIDDEN           # 403 — 다른 수신자 목록 조회 시도
            - OPERATOR_ONLY                 # 403 — 운영자 전용 작업을 일반 USER 가 시도
            - INVALID_STATE                 # 409 — 도메인 상태 전제 위반 (예: PENDING 알림에 retry)
            - OPTIMISTIC_LOCK_CONFLICT      # 409 — 동시 갱신 충돌 (재시도 가능)
            - VALIDATION_FAILED             # 400 — 요청 본문 / 파라미터 검증 실패
            - BAD_REQUEST                   # 400 — IllegalArgumentException 일반
            - NOTIFICATION_DATA_INCONSISTENCY  # 500 — outbox/notification 매핑 깨짐 (운영 alert)
            - INTERNAL_ERROR                # 500 — 분류되지 않은 내부 오류
        message:
          type: string